added exceptions

+++ Module 3/Day 5/CustomerProduct/src/main/java/org/northernarc/customerproduct/service/CustomerServiceDaoImpl.java
@@ -25,6 +25,9 @@ public class CustomerServiceDaoImpl implements CustomerServiceDao {
 
     @Override
     public void deleteById(Integer id) {
+        if (!customerRepository.existsById(id)) {
+            throw new CustomerNotFound("no customer found with id " + id);
+        }
         customerRepository.deleteById(id);
     }
